requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
redis
python-dotenv
httpx
pytest
pytest-asyncio
